Add console cron command for cronjobs
inspired by the slim 4 skeleton (https://odan.github.io/slim4-skeleton/console.html)
