+++ src/Domain/Main/MainService.php
@@ -4,197 +4,17 @@ namespace App\Domain\Main;
 
 use App\Domain\Service;
 use Psr\Log\LoggerInterface;
-use App\Domain\Base\Settings;
 use App\Domain\Base\CurrentUser;
-use Dubture\Monolog\Reader\LogReader;
-use App\Domain\Finances\Recurring\RecurringFinanceEntryCreator;
-use App\Domain\Finances\FinanceStatsMonthlyMailService;
-use App\Domain\Board\Card\CardMailService;
-use App\Domain\User\Token\TokenService;
 use App\Application\Payload\Payload;
-use App\Domain\Settings\SettingsMapper;
 use Slim\Csrf\Guard as CSRF;
-use App\Domain\Splitbill\RecurringBill\RecurringBillEntryCreator;
 
 class MainService extends Service {
 
-    private $settings;
-    protected $settings_mapper;
-    protected $finances_entry_creator;
-    protected $finance_stats_monthly_mail_service;
-    protected $card_mail_service;
-    protected $token_service;
     private $csrf;
-    private $bill_entry_creator;
 
-    public function __construct(LoggerInterface $logger,
-            CurrentUser $user,
-            Settings $settings,
-            SettingsMapper $settings_mapper,
-            RecurringFinanceEntryCreator $finances_entry_creator,
-            FinanceStatsMonthlyMailService $finance_stats_monthly_mail_service,
-            CardMailService $card_mail_service,
-            TokenService $token_service,
-            CSRF $csrf,
-            RecurringBillEntryCreator $bill_entry_creator) {
+    public function __construct(LoggerInterface $logger, CurrentUser $user, CSRF $csrf) {
         parent::__construct($logger, $user);
-
-        $this->settings = $settings;
-        $this->settings_mapper = $settings_mapper;
-        $this->finances_entry_creator = $finances_entry_creator;
-        $this->finance_stats_monthly_mail_service = $finance_stats_monthly_mail_service;
-        $this->card_mail_service = $card_mail_service;
-        $this->token_service = $token_service;
         $this->csrf = $csrf;
-        $this->bill_entry_creator = $bill_entry_creator;
-    }
-
-    public function cron(): Payload {
-        $this->logger->info('Running CRON');
-
-        $lastRunRecurring = $this->settings_mapper->getSetting("lastRunRecurring");
-        $lastRunFinanceSummary = $this->settings_mapper->getSetting("lastRunFinanceSummary");
-        $lastRunCardReminder = $this->settings_mapper->getSetting("lastRunCardReminder");
-        $lastRunRecurringSplitbills = $this->settings_mapper->getSetting("lastRunRecurringSplitbills");
-
-        $date = new \DateTime('now');
-
-        // Update recurring finances @ 06:00
-        if ($date->format("H") === "06" && $lastRunRecurring->getDayDiff() > 0) {
-            $this->logger->notice('CRON - Update Finances');
-
-            $this->finances_entry_creator->update();
-            $this->settings_mapper->updateLastRun("lastRunRecurring");
-        }
-
-        // Is first of month @ 08:00? Send Finance Summary
-        if ($date->format("d") === "01" && $date->format("H") === "08" && $lastRunFinanceSummary->getDayDiff() > 0) {
-            $this->logger->notice('CRON - Send Finance Summary');
-
-            $this->finance_stats_monthly_mail_service->sendSummary();
-            $this->settings_mapper->updateLastRun("lastRunFinanceSummary");
-        }
-
-        // card reminder @ 09:00
-        if ($date->format("H") === "09" && $lastRunCardReminder->getDayDiff() > 0) {
-            $this->logger->notice('CRON - Send Card Reminder');
-
-            $this->card_mail_service->sendReminder();
-            $this->settings_mapper->updateLastRun("lastRunCardReminder");
-
-//            $this->token_service->deleteOldTokens();
-        }
-
-        // Update recurring splitted bills @ 06:00
-        if ($date->format("H") === "06" && $lastRunRecurringSplitbills->getDayDiff() > 0) {
-            $this->logger->notice('CRON - Update Splitted Bills');
-
-            $this->bill_entry_creator->update();
-            $this->settings_mapper->updateLastRun("lastRunRecurringSplitbills");
-        }
-
-        $response_data = ['result' => 'success'];
-
-        return new Payload(Payload::$RESULT_JSON, $response_data);
-    }
-
-    public function getLogfileOverview($days) {
-        return new Payload(Payload::$RESULT_HTML, ["days" => $days]);
-    }
-
-    public function getLogfile($days) {
-        $file = new \SplFileObject($this->settings->all()['logger']['path'], 'r');
-
-        $pattern = '/\[(?P<date>.*)\] (?P<logger>[\w\-]+).(?P<level>\w+): (?P<message>[^\{]+) (?P<context>[\[\{].*[\]\}]) (?P<extra>[\[\{].*[\]\}])/';
-
-        $logfile = array();
-
-        while (!$file->eof()) {
-
-            $logline = $file->current();
-            $log = $this->parseLogLine($logline, $days, $pattern);
-
-            // do not show datatable queries
-            if (!empty($log)) {
-
-                $line = array();
-
-                /**
-                 * Get Username
-                 */
-                $user = "";
-                if (array_key_exists("user", $log["extra"]) && !is_null($log["extra"]["user"])) {
-                    $user = $log["extra"]["user"];
-                } elseif (array_key_exists("login", $log["context"])) {
-                    $user = $log["context"]["login"];
-                }
-
-                // Remove surrounding "Array ( ... )'
-                $regex = '/^Array\s*\((.*)?\)\s*$/s';
-
-                $line["user"] = $user;
-                $line["date"] = $log['date']->format('Y-m-d H:i:s');
-                $line["level"] = $log['level'];
-                $line["message"] = $log['message'];
-                $line["extra"] = preg_replace($regex, '$1', print_r($log['extra'], true));
-                $line["context"] = !empty($log["context"]) ? preg_replace($regex, '$1', print_r($log['context'], true)) : null;
-                $line["url"] = array_key_exists("url", $log['extra']) ? $log['extra']["url"] : null;
-                $line["hide"] = strpos($line["url"], "datatable=1") == 0 ? false : true;
-
-                if (strlen($line["url"]) > 100 && array_key_exists("query", $log['extra'])) {
-                    $line["url"] = str_replace($log['extra']["query"], "...", $line["url"]);
-                }
-
-                $line["query"] = array_key_exists("query", $log['extra']) ? $log['extra']["query"] : null;
-
-                array_push($logfile, $line);
-            }
-            $file->next();
-        }
-
-        return new Payload(Payload::$RESULT_HTML, ["logfile" => $logfile]);
-    }
-
-    private function parseLogLine($log, $days, $pattern) {
-        if (!is_string($log) || strlen($log) === 0) {
-            return array();
-        }
-
-        $data = [];
-        preg_match($pattern, $log, $data);
-
-        if (!isset($data['date'])) {
-            return array();
-        }
-
-        $date = \DateTime::createFromFormat('Y-m-d H:i:s', $data['date']);
-        // try monolog 2 date format
-        if (!$date) {
-            $date = \DateTime::createFromFormat('Y-m-d\TH:i:s.uP', $data['date']);
-        }
-
-        $array = array(
-            'date' => $date,
-            'logger' => $data['logger'],
-            'level' => $data['level'],
-            'message' => $data['message'],
-            'context' => json_decode($data['context'], true),
-            'extra' => json_decode($data['extra'], true)
-        );
-
-        if (0 === $days) {
-            return $array;
-        }
-
-        if (isset($date) && $date instanceof \DateTime) {
-            $d2 = new \DateTime('now');
-
-            if ($date->diff($d2)->days < $days) {
-                return $array;
-            } else {
-                return array();
-            }
-        }
     }
 
     public function getCSRFTokens($count) {
